+++ src/modules/gateway/Anonymise.ts
@@ -1,11 +1,12 @@
 import fs from "fs";
 import path from "path";
+import { DaikinCloudDevice } from "../../daikin-cloud";
 import { getNewConfigDir } from "../paths";
 
-function anonymise(dev: any, value: string | undefined) {
-	let fileName = value ?? dev.getId();
-
-	let data = recurse(dev)
+function anonymise(dev: DaikinCloudDevice, value: string | undefined) {
+	const fileName = value ?? dev.getId();
+	const snapshot = structuredClone(dev.getDescription());
+	const data = recurse(snapshot);
 	const configFolder = getNewConfigDir()
 	const configFile = path.join(configFolder, fileName + '.json')